Cancel image

+++ app/lib/KuraStar/Storage/Article/EloquentArticleRepository.php
@@ -51,7 +51,13 @@ class EloquentArticleRepository implements ArticleRepository{
 			if(fopen($path, "w")){
 				if(copy($input['imageUrl'], $path)){
 					$picture = Article::where('CURATION_ID', '=', $input['cur_id'])->first();
-					if(unlink(public_path().$picture['CURATION_IMAGE'])){
+					if($picture['CURATION_IMAGE'] != ""){
+						if(unlink(public_path().$picture['CURATION_IMAGE'])){
+							$article = Article::where('CURATION_ID', '=', $input['cur_id'])
+								->update(['CURATION_IMAGE' => $file]);
+						}
+					}
+					else{
 						$article = Article::where('CURATION_ID', '=', $input['cur_id'])
 								->update(['CURATION_IMAGE' => $file]);
 					}
